backend/requirements.txt
fastapi
uvicorn
pycryptodome
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv